Was trying out minecraft mappings but reverted to before, will create a new branch for that.
