--- remappedSrc/me/av306/keybindsgaloreplus/customdata/KeybindData.java
@@ -1,10 +1,0 @@
-package me.av306.keybindsgaloreplus.customdata;
-
-public class KeybindData
-{
-    public String displayName = null;
-
-    public Integer sectorColor = null;
-
-    public boolean hideCategory = false;
-}
