--- remappedSrc/me/av306/keybindsgaloreplus/mixin/MinecraftClientAccessor.java
@@ -1,11 +1,0 @@
-package me.av306.keybindsgaloreplus.mixin;
-
-import net.minecraft.client.MinecraftClient;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
-
-@Mixin( MinecraftClient.class )
-public interface MinecraftClientAccessor
-{
-    @Accessor void setAttackCooldown( int attackCooldown );
-}
